refactor: 태그, 카테고리, 공동 생활습관 타입을 common 타입으로 분리
